Ensure that autodrain() gets called every time
