+++ lib/stream/xlsx/hyperlink-reader.js
@@ -1,4 +1,5 @@
 const {EventEmitter} = require('events');
+const iterateStream = require('../../utils/iterate-stream');
 const parseSax = require('../../utils/parse-sax');
 
 const Enums = require('../../doc/enums');
@@ -44,7 +45,7 @@ class HyperlinkReader extends EventEmitter {
     }
 
     try {
-      for await (const events of parseSax(entry)) {
+      for await (const events of parseSax(iterateStream(entry))) {
         for (const {eventType, value} of events) {
           if (eventType === 'opentag') {
             const node = value;
