+++ lib/stream/xlsx/workbook-reader.js
@@ -4,6 +4,7 @@ const {PassThrough, Readable} = require('readable-stream');
 const nodeStream = require('stream');
 const unzip = require('unzipper');
 const tmp = require('tmp');
+const iterateStream = require('../../utils/iterate-stream');
 const parseSax = require('../../utils/parse-sax');
 
 const StyleManager = require('../../xlsx/xform/style/styles-xform');
@@ -79,17 +80,18 @@ class WorkbookReader extends EventEmitter {
   async *parse(input, options) {
     if (options) this.options = options;
     const stream = (this.stream = this._getStream(input || this.input));
+    const zip = unzip.Parse({forceStream: true});
+    stream.pipe(zip);
 
     // worksheets, deferred for parsing after shared strings reading
     const waitingWorkSheets = [];
 
-    for await (const entry of this._iterateZip(stream)) {
+    for await (const entry of iterateStream(zip)) {
       let match;
       let sheetNo;
       switch (entry.path) {
         case '_rels/.rels':
         case 'xl/_rels/workbook.xml.rels':
-          entry.autodrain();
           break;
         case 'xl/workbook.xml':
           await this._parseWorkbook(entry);
@@ -127,11 +129,10 @@ class WorkbookReader extends EventEmitter {
             match = entry.path.match(/xl\/worksheets\/_rels\/sheet(\d+)[.]xml.rels/);
             sheetNo = match[1];
             yield* this._parseHyperlinks(entry, sheetNo);
-          } else {
-            entry.autodrain();
           }
           break;
       }
+      entry.autodrain();
     }
 
     for (const {sheetNo, path, tempFileCleanupCallback} of waitingWorkSheets) {
@@ -154,7 +155,7 @@ class WorkbookReader extends EventEmitter {
 
   async _parseWorkbook(entry) {
     this._emitEntry({type: 'workbook'});
-    await this.properties.parseStream(entry);
+    await this.properties.parseStream(iterateStream(entry));
   }
 
   async *_parseSharedStrings(entry) {
@@ -166,14 +167,13 @@ class WorkbookReader extends EventEmitter {
       case 'emit':
         break;
       default:
-        entry.autodrain();
         return;
     }
 
     let inT = false;
     let text = null;
     let index = 0;
-    for await (const events of parseSax(entry)) {
+    for await (const events of parseSax(iterateStream(entry))) {
       for (const {eventType, value} of events) {
         if (eventType === 'opentag') {
           const node = value;
@@ -202,9 +202,7 @@ class WorkbookReader extends EventEmitter {
     this._emitEntry({type: 'styles'});
     if (this.options.styles === 'cache') {
       this.styles = new StyleManager();
-      await this.styles.parseStream(entry);
-    } else {
-      entry.autodrain();
+      await this.styles.parseStream(iterateStream(entry));
     }
   }
 
@@ -223,44 +221,6 @@ class WorkbookReader extends EventEmitter {
       yield {eventType: 'hyperlinks', value: hyperlinksReader};
     }
   }
-
-  async *_iterateZip(stream) {
-    const zip = unzip.Parse();
-    stream.pipe(zip);
-
-    const entries = [];
-    zip.on('entry', entry => entries.push(entry));
-
-    let finished = false;
-    zip.on('finish', () => (finished = true));
-    const zipFinishedPromise = once(zip, 'finish');
-
-    while (!finished || entries.length > 0) {
-      if (entries.length === 0) {
-        // eslint-disable-next-line no-await-in-loop
-        await Promise.race([once(zip, 'entry'), zipFinishedPromise]);
-      } else {
-        const entry = entries.shift();
-        yield entry;
-        entry.autodrain();
-      }
-    }
-  }
-}
-
-function once(eventEmitter, type) {
-  // TODO: Use require('events').once when node v10 is dropped
-  return new Promise(resolve => {
-    let fired = false;
-    const handler = () => {
-      if (!fired) {
-        fired = true;
-        eventEmitter.removeListener(type, handler);
-        resolve();
-      }
-    };
-    eventEmitter.addListener(type, handler);
-  });
 }
 
 // for reference - these are the valid values for options
