+++ lib/stream/xlsx/worksheet-reader.js
@@ -1,4 +1,5 @@
 const {EventEmitter} = require('events');
+const iterateStream = require('../../utils/iterate-stream');
 const parseSax = require('../../utils/parse-sax');
 
 const _ = require('../../utils/under-dash');
@@ -154,7 +155,7 @@ class WorksheetReader extends EventEmitter {
     let row = null;
     let c = null;
     let current = null;
-    for await (const events of parseSax(entry)) {
+    for await (const events of parseSax(iterateStream(entry))) {
       const worksheetEvents = [];
       for (const {eventType, value} of events) {
         if (eventType === 'opentag') {
