+++ lib/xlsx/xform/base-xform.js
@@ -56,7 +56,7 @@ class BaseXform {
     this.model = Object.assign(this.model || {}, obj);
   }
 
-  async parse(saxParser, stream) {
+  async parse(saxParser) {
     for await (const events of saxParser) {
       for (const {eventType, value} of events) {
         if (eventType === 'opentag') {
@@ -65,7 +65,6 @@ class BaseXform {
           this.parseText(value);
         } else if (eventType === 'closetag') {
           if (!this.parseClose(value.name)) {
-            if (stream) stream.emit('finish'); // needed for unzipper to continue to the next entry
             return this.model;
           }
         }
@@ -75,7 +74,7 @@ class BaseXform {
   }
 
   async parseStream(stream) {
-    return this.parse(parseSax(stream), stream);
+    return this.parse(parseSax(stream));
   }
 
   get xml() {
